requirements.txt
# This file is @generated by PDM.
# Please do not edit it manually.

aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.2.0
alembic==1.10.3
anyio==3.6.1
arrow==1.2.3
async-lru==2.0.2
async-timeout==4.0.2
asyncio==3.4.3
asyncodbc==0.1.1
attrs==21.4.0
azure-common==1.1.28
azure-core==1.26.4
azure-cosmos==4.3.1
azure-data-tables==12.4.2
azure-eventgrid==4.9.1
azure-functions==1.13.3
azure-identity==1.12.0
azure-keyvault-secrets==4.7.0
azure-servicebus==7.11.3
azure-storage-blob==12.15.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cryptography==37.0.4
deepdiff==6.3.0
frozenlist==1.3.0
greenlet==2.0.2
h11==0.12.0
h2==4.1.0
hpack==4.0.0
httpcore==0.15.0
httpx==0.24.0
hyperframe==6.0.1
idna==3.3
inflection==0.5.1
isodate==0.6.1
Mako==1.2.4
MarkupSafe==2.1.2
msal==1.18.0
msal-extensions==1.0.0
msrest==0.7.1
multidict==6.0.2
mygeotab==0.8.8
numpy==1.26.0
oauthlib==3.2.0
ordered-set==4.1.0
orjson==3.8.10
pandas==2.1.1
portalocker==2.5.1
pycparser==2.21
PyJWT==2.4.0
pyodbc==4.0.34
python-dateutil==2.8.2
python-rapidjson==1.9
pytz==2022.7.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.2.0
sqlalchemy==2.0.9
sqlanydb==1.0.13
twilio==8.0.0
typing-extensions==4.3.0
tzdata==2023.3
urllib3==1.26.11
yarl==1.7.2
